Changed how attraction force works in the graph
Added a method moveTowards(target, force), where nodes move towards a target but if the force would push them past it then it instead sets their position to the target position. Also changed it so that attractive forces sum up and move towards the average position of all neighbours instead of just on at a time, should be more accurate?

Additionally, added a new movement rule for the terrain where agents just teleport to and from work, to make the simulation run faster.