Graph Drawing
Fixed force-directed graph drawing, and moved the draw method to the update method in graph renderer. Update is frame-dependent so not too intensive, and for a low number of nodes it only took about 50ms (More expensive with more nodes obvs). Small issue is that the setPosition method in the nodescript doesn't factor in the rigid body so it avoids collision detection, meaining that nodes end up instide of each other :(.