Ideal Graph & JSON stuff
Added the support to compile the game to C++, though it doesn't work (Gameobjects don't seem to be instantiated). Made the MinHeap convertable to JSON (Along with agent via a built in function). Added a toggle for the graph so that it can be turned into an "Ideal graph" (as I called it). This means that edges are formed on PURELY personality similarity. Toggling the option will redraw the graph. Currently, it only calculates the clustering coefficient of the ideal graph.